requirements.txt
streamlit
openai
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
